test(web): split dashboard test suites [OPE-398]
